Warning en compilacion resueltos

+++ DeathTokenBase/DeathToken/Menu.cpp
@@ -4,10 +4,10 @@
 
 Menu::Menu(Game* game) : GameState(game), texture(game->getTexture(BACKGROUND)) {
 	//Widht, height, position baccarat button
-	int wBut = Game::WIN_WIDTH / 6.8, hBut = Game::WIN_HEIGHT / 4.5,
+	double wBut = Game::WIN_WIDTH / 6.8, hBut = Game::WIN_HEIGHT / 4.5,
 		xBut = Game::WIN_WIDTH / 4 - Game::WIN_WIDTH / 8, yBut = Game::WIN_HEIGHT / 4 + Game::WIN_HEIGHT / 12.2;
 	//Baccarat button
-	baccarat = new Button(this, xBut, yBut, wBut, hBut, game->getTexture(BACCARATBUT));
+	baccarat = new Button(this,(int) xBut, (int)yBut, (int)wBut, (int)hBut, game->getTexture(BACCARATBUT));
 	addObjects(baccarat);
 	addEventListener(baccarat);
 	baccarat->connect([this]() { gameChanger(new Baccarat(getGame())); });
@@ -18,10 +18,10 @@ Menu::Menu(Game* game) : GameState(game), texture(game->getTexture(BACKGROUND)) 
 	slots->connect([this]() { gameChanger(new Slots(getGame())); });
 
 	//Widht, height, position marbles button
-	wBut = Game::WIN_WIDTH / 5.2; hBut = Game::WIN_HEIGHT / 4;
-	xBut = Game::WIN_WIDTH * 3 / 4 - Game::WIN_WIDTH / 25; yBut = Game::WIN_HEIGHT / 4 + Game::WIN_HEIGHT / 25;
+	wBut = Game::WIN_WIDTH / 5.2; hBut = Game::WIN_HEIGHT / 4.0;
+	xBut = Game::WIN_WIDTH * 3.0 / 4.0 - Game::WIN_WIDTH / 25.0; yBut = Game::WIN_HEIGHT / 4.0 + Game::WIN_HEIGHT / 25.0;
 	//Marbles button
-	marbles = new Button(this, xBut, yBut, wBut, hBut, game->getTexture(CANICASBUT));
+	marbles = new Button(this, (int)xBut, (int)yBut, (int)wBut, (int)hBut, game->getTexture(CANICASBUT));
 	addObjects(marbles);
 	addEventListener(marbles);
 	marbles->connect([this]() { gameChanger(new Marbles(getGame())); });
